Tile print completed. User-input movement (but not other) commands completed.
Note:
CLion cannot debug Ncurses (crashes on initscr) without a lot of workaround magic which I have not yet done, but it can run it fine. This is not a problem in Pyrite.
CLion does not print the map correctly in that if it prints the rightmost column, all prints will be on one line. This is not a problem in Pyrite.
CLion ncurses input incorrectly requires an enter after a key. This is not a problem in Pyrite.